+++ src/utils.js
@@ -6,7 +6,7 @@ const ignore = require('ignore');
 const MIGRATION_FOLDER = 'migration';
 const MIGRATION_TEMPLATES_FOLDER = MIGRATION_FOLDER + '/template';
 const PROJECT_INFO_FOLDER = MIGRATION_FOLDER + '/project_info';
-const CORDOVA_VERSION = '7.1.0';
+const CORDOVA_VERSION = '9.0.0';
 
 
 let filterIgnoreFiles = function(files, ignoreList, removeBasePath = false) {