+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 CHANGELOG
 ====
 
+v3.2.11
+----
+* Update dependencies to resolve circular dependency warnings
+
+v3.2.10
+----
+* Update Cordova 10 Blank Template for Project Migration Command
+
+v3.2.9
+----
+* Fixed Resource Permission for Localkit on Non-Windows
+
+v3.2.8
+----
+* Upgrade default blank template to the latest version
+* Update default cordova version to 9.0.0
+
 v3.2.7
 ----
 * Fixes to support older version of Nodejs (CERT_HAS_EXPIRED).